+++ Changelog.md
@@ -45,6 +45,7 @@
   - [#269](../../issues/269) - Show underscore of header text on RibbonTabItem
   - [#272](../../issues/272) - Changing RibbonThemeColorBrush does not change background of ItemsPanel in Backstage
   - [#274](../../issues/274) - RadioButton Icon and LargeIcon
+  - [#280](../../issues/280) - Keytips of the Ribbon overlay StartScreen
   - OpenBackstage command was not acting on the correct backstage in a multiple backstage scenario (thanks to @maurosampietro)
 
 - ### Enhancements